revert submitMethod back into method-common.js
