+++ src/workflows-app/utils/method-common.js
@@ -1,5 +1,10 @@
 import { Ajax } from 'src/libs/ajax';
+import { Apps } from 'src/libs/ajax/leonardo/Apps';
+import * as Nav from 'src/libs/nav';
+import { notify } from 'src/libs/notifications';
+import { getUser } from 'src/libs/state';
 import * as Utils from 'src/libs/utils';
+import { resolveRunningCromwellAppUrl } from 'src/libs/workflows-app-utils';
 
 const MethodSource = Object.freeze({
   GitHub: 'GitHub',
@@ -8,6 +13,37 @@ const MethodSource = Object.freeze({
 
 const Covid19Methods = ['fetch_sra_to_bam', 'assemble_refbased', 'sarscov2_nextstrain'];
 export const isCovid19Method = (methodName) => Covid19Methods.includes(methodName);
+
+export const submitMethod = async (signal, onDismiss, method, workspace) => {
+  const namespace = await workspace.workspace.namespace;
+  try {
+    const cbasUrl = (
+      await Apps(signal)
+        .listAppsV2(workspace.workspace.workspaceId)
+        .then((apps) => resolveRunningCromwellAppUrl(apps, getUser()?.email))
+    ).cbasUrl;
+
+    if (cbasUrl) {
+      const methodPayload = {
+        method_name: method.method_name,
+        method_description: method.method_description,
+        method_source: method.method_source,
+        method_version: method.method_version,
+        method_url: method.method_url,
+      };
+      const methodObject = await Ajax(signal).Cbas.methods.post(cbasUrl, methodPayload);
+      onDismiss();
+      Nav.goToPath('workspace-workflows-app-submission-config', {
+        name: workspace.workspace.name,
+        namespace,
+        methodId: methodObject.method_id,
+      });
+    }
+  } catch (error) {
+    notify('error', 'Error creating new method', { detail: error instanceof Response ? await error.text() : error });
+    onDismiss();
+  }
+};
 
 export const convertToRawUrl = (methodPath, methodVersion, methodSource) => {
   return Utils.cond(
